requirements.txt
tensorflow 
pandas
gdown
dvc
mlflow
notebook
numpy
matplotlib
seaborn
python-box
pyYAML
tqdm
ensure
joblib
typeS-PyYAML
scipy
Flask
Flask-Cors
dagshub

-e .